Api finalizada na unha para aperfeiçoar

+++ services/product.service.js
@@ -1,8 +1,9 @@
 import ProductRepository from "../repositories/product.repository.js";
 import SupplierRepository from "../repositories/supplier.repository.js";
+import SaleRepository from "../repositories/sale.repository.js";
 
 async function createProduct(product) {
-  if (await SupplierRepository.getSupplier(product.supplier_id)) {
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
     return await ProductRepository.insertProduct(product);
   }
   throw new Error("O supplier_id informado não existe.");
@@ -23,24 +24,24 @@ async function getProduct(id) {
 }
 
 async function deleteProduct(id) {
-  if (await ProductRepository.getProduct(id)) {
-    return await ProductRepository.deleteProduct(id);
+  const sales = await SaleRepository.getSalesByProductId(id);
+  if (sales) {
+    throw new Error(
+      "Não é possível excluir o product pois ele já foi vendido."
+    );
+  } else {
+    if (await ProductRepository.getProduct(id)) {
+      return await ProductRepository.deleteProduct(id);
+    }
+    throw new Error("O product informado não existe");
   }
-  throw new Error("O product informado não existe");
 }
 
 async function updateProduct(product) {
-  let error = "";
-  if (!(await SupplierRepository.getSupplier(product.supplier_id))) {
-    error += "O supplier_id informado não existe ";
+  if (await SupplierRepository.getSupplier(product.supplierId)) {
+    return await ProductRepository.updateProduct(product);
   }
-  if (!(await ProductRepository.getProduct(product.product_id))) {
-    error += "O product_id informado não existe";
-  }
-  if (error) {
-    throw new Error(error);
-  }
-  return await ProductRepository.insertProduct(product);
+  throw new Error("O supplier_id informado não existe.");
 }
 
 export default {
